requirements.txt
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.109.0
greenlet==3.3.1
h11==0.16.0
python-dotenv==1.2.1
SQLAlchemy==2.0.46
sqlmodel==0.0.14
typing_extensions==4.15.0
tzdata==2025.3
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.2.0
numpy==1.26.0
shap
joblib
xgboost
scikit-learn
networkx
celery[redis]
openai>=1.10.0

# Testing & Quality Assurance
pytest
pytest-cov
httpx

# Observability
sentry-sdk[fastapi]
loguru

# Coding Standards
ruff
pre-commit

pinecone[grpc]
